Update mysql-5* services to be ARM ready
The Docker compose file has been updated. This includes running the step-ca container as root, adding error handling and certificate ownership change to the script running at startup, and removing the specification of a user for the mysql-5 service. It also includes dynamically injecting the database port number and correctly setting the PMA_ABSOLUTE_URI for phpMyAdmin.